requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
asyncio>=3.4.3
requests>=2.28.0
pillow>=9.0.0
pytz>=2022.1
